👷Extract logic @ home to useGetUserInfo hook
