+++ src/Pages/Private/Home.jsx
@@ -2,25 +2,10 @@ import CardsArray from 'Components/Elements/CardsArray'
 import HomeActions from 'Components/Home/HomeActions'
 import HomeFunds from 'Components/Home/HomeFunds'
 import HomeHeader from 'Components/Home/HomeHeader'
-import { useEffect, useState } from 'react'
-import UserInfo from 'Services/User'
+import useGetUserInfo from 'Hooks/useGetUserInfo'
 
 const Home = () => {
-  const [user, setUser] = useState({})
-
-  useEffect(() => {
-    const getData = async () => {
-      try {
-        const res = await UserInfo.getUserInfo()
-        setUser(res.data)
-      } catch (err) {
-        console.log(err)
-      }
-    }
-    getData()
-
-    // get paymethds
-  }, [])
+  const { user } = useGetUserInfo()
 
   return (
     <>
